Updated product code type to long

+++ Server/src/main/java/app/types/Product.java
@@ -1,4 +1,4 @@
 package app.types;
 
-public record Product(int code, String name, double price, int supplierId) implements DataType {
+public record Product(long code, String name, double price, int supplierId) implements DataType {
 }
